audio_transcription_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.9.2

openai-whisper==20240930
ffmpeg-python==0.2.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
--extra-index-url https://download.pytorch.org/whl/cu124

numpy==1.26.0
aiohttp==3.13.1
psutil==5.9.8